[Fix/#500] 콕 찌르기 유저 프로필 사진 아웃라인에 색상 적용

+++ feature/poke/src/main/java/org/sopt/official/feature/poke/poke_user_recycler_view/ItemDecorationDivider.kt
@@ -4,10 +4,9 @@ import android.graphics.Canvas
 import android.graphics.Paint
 import androidx.recyclerview.widget.RecyclerView
 
-
 class ItemDecorationDivider(
     color: Int,
-    private val height: Float,
+    private val height: Int,
 ) : RecyclerView.ItemDecoration() {
 
     private val paint = Paint()
